requirements.txt
numpy==1.19.2
torch==1.8.1
torchvision==0.9.1
pyro-ppl==1.8.4
scikit-learn==0.23.1